requirement.txt
django
djangorestframework
Pillow
drf-spectacular
django-extensions